+++ src/components/AppEssentialLink.vue
@@ -21,7 +21,10 @@ withDefaults(defineProps<{
       v-if="icon"
       avatar
     >
-      <q-icon :name="icon" />
+      <q-icon
+        :name="icon"
+        size="md"
+      />
     </q-item-section>
 
     <q-item-section>